# perf(zod): prefer to use `partial` to make all properties optional