--- a/packages/rspack/src/builtin-loader/swc/collectTypeScriptInfo.ts
+++ b/packages/rspack/src/builtin-loader/swc/collectTypeScriptInfo.ts
@@ -1,9 +1,3 @@
-import * as z from "zod/v4";
-
 export type CollectTypeScriptInfoOptions = {
 	typeExports?: boolean;
 };
-
-export const ZodSwcCollectTypeScriptInfo = z.strictObject({
-	typeExports: z.boolean().optional()
-}) satisfies z.ZodType<CollectTypeScriptInfoOptions>;