--- a/packages/rspack/src/builtin-loader/swc/pluginImport.ts
+++ b/packages/rspack/src/builtin-loader/swc/pluginImport.ts
@@ -1,5 +1,3 @@
-import * as z from "zod/v4";
-
 type RawStyleConfig = {
 	styleLibraryDirectory?: string;
 	custom?: string;
@@ -33,21 +31,6 @@ type PluginImportConfig = {
 };
 
 type PluginImportOptions = PluginImportConfig[];
-
-export const ZodSwcPluginImportConfig = z
-	.strictObject({
-		libraryName: z.string(),
-		libraryDirectory: z.string().optional(),
-		customName: z.string().optional(),
-		customStyleName: z.string().optional(),
-		style: z.string().or(z.boolean()).optional(),
-		styleLibraryDirectory: z.string().optional(),
-		camelToDashComponentName: z.boolean().optional(),
-		transformToDefaultImport: z.boolean().optional(),
-		ignoreEsComponent: z.string().array().optional(),
-		ignoreStyleComponent: z.string().array().optional()
-	})
-	.array() satisfies z.ZodType<PluginImportOptions>;
 
 function isObject(val: any): boolean {
 	return Object.prototype.toString.call(val) === "[object Object]";