## Summary

Prefer to use `partial` to make all properties optional, this makes the code more readable and can reduce bundle size.

> https://zod.dev/api?id=partial

## Bundle size

- before:

```txt
File (cjs0)     Size       
dist/index.js   850.5 kB
```

- after:

```txt
File (cjs0)     Size       
dist/index.js   843.9 kB
```

## Runtime performance

`optional()` and `partial()` have similar performance.

```bash
hyperfine --warmup 5 --runs 20  'node ./optional.js' 'node ./partial.js'
Benchmark 1: node ./optional.js
  Time (mean ± σ):     201.0 ms ±  14.5 ms    [User: 332.4 ms, System: 25.3 ms]
  Range (min … max):   190.1 ms … 242.9 ms    20 runs
 
Benchmark 2: node ./partial.js
  Time (mean ± σ):     198.0 ms ±   4.1 ms    [User: 332.0 ms, System: 25.0 ms]
  Range (min … max):   192.7 ms … 211.5 ms    20 runs
 
Summary
  node ./partial.js ran
    1.02 ± 0.08 times faster than node ./optional.js
```

## Checklist

<!--- Check and mark with an "x" -->

- [ ] Tests updated (or not required).
- [ ] Documentation updated (or not required).